backtraceMAC - fix errors in MAC algorithm
This commit introduces two checks in the MAC realignment algorithm
in order to alleviate Issue #153 ("MAC realignment algorithm damages hits",
https://github.com/soedinglab/hh-suite/issues/153):

1. If MAC hit is significantly shorter than the Viterbi
hit in terms of number of matched columns then the original Viterbi hit
is restored and an appropriate warning is issued.
This happens when MAC algorithm generates several disjoint local
hits or when MAC hit is shorter due to weak score at the start/end
regions of the Viterbi hit.
The threshold on the MAC hit length is defined by a new
option, -mac_min_length (default=0.9, i.e. at least 90% of number of
matched columns in Viterbi hit must be in the corresponding MAC hit).

There is no obvious way to deal with hit split in the current
MAC algorithm aproach, it is expected that the algorithm should
find the same Viterbi hit albeit improved and extended at its borders.
Usually the MAC hit restores its integrity as soon as the -mact
option is decreased. For example, in the test case described in the
issue report one could find the proper MAC hit if "-mact 0.33" is used
(instead of the default "-mact 0.35" value).
Therefore the warning also suggests decreasing -mact option as a viable
solution if MAC alignment is still needed.

Note: It would be nice to have support for "-mact auto" option
that forced the MAC algorithm to decrease -mact to the max. value
that restores the hit integrity (i.e. match the original Viterbi
hit to the degree defined by -mac_min_length option).

2. MAC realignment algorithm also can find a local hit that is
completely disjoint from the original Viterbi hit.
This possibility follows from the fact that viterbi_matrix
used in MAC DP algorithm is masked (see maskViterbiAlignment())
in a way that allows extension of the hit.
This commit adds a check that the found MAC hit has intersection
with Viterbi hit in both, query and target, coordinate spaces.
If the check fails, the alternative MAC hit that satisfies this
condition is sought (in decreasing score order).
When no valid MAC hit is found, the original Viterbi hit is restored
and an appropriate warning is printed.
This error was also observed in the reported issue, see the hit #4
(Score=11.3) that was mapped erroneously to the second part of
main hit, template columns [261; 357].

The alternative MAC hits' positions are collected in macAlgorithm()
procedure and later used in backtraceMAC() procedure.